+++ frontend/src/app/shared/models/account.ts
@@ -1,0 +1,8 @@
+export class Account {
+    Account: string;
+    EDRPO: string;
+    Id: number;
+    MFO: number;
+    NAME: string;
+    Narrative: string;
+}
