+++ frontend/src/app/shared/models/index.ts
@@ -1,2 +1,3 @@
 export * from './payment';
 export * from './dbf';
+export * from './account';
